+++ Source/EventFlow.Tests/UnitTests/Extensions/StringExtensionsTests.cs
@@ -20,7 +20,8 @@
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// 
+//
+
 using EventFlow.Extensions;
 using EventFlow.TestHelpers;
 using FluentAssertions;
@@ -37,6 +38,13 @@ namespace EventFlow.Tests.UnitTests.Extensions
         public void ToSlug(string input, string expected)
         {
             input.ToSlug().Should().Be(expected);
+        }
+
+        [TestCase("magic number", "53033b46d5c023c6e99d734db2215d580020fa9c4091cd7edf07fa48abaf058e")]
+        [TestCase("a big secret", "3d530bd404d0496c073b8aa927ca8911911c81b4939510af25533abaa6d323b7")]
+        public void ToSha256(string input, string expected)
+        {
+            input.ToSha256().Should().Be(expected);
         }
     }
 }