+++ Source/EventFlow/Core/ITransientFaultHandler.cs
@@ -32,6 +32,11 @@ namespace EventFlow.Core
     {
         void ConfigureRetryStrategy(Action<TRetryStrategy> configureStrategy);
 
+        Task TryAsync(
+            Func<CancellationToken, Task> action,
+            Label label,
+            CancellationToken cancellationToken);
+
         Task<T> TryAsync<T>(
             Func<CancellationToken, Task<T>> action,
             Label label,