+++ Source/EventFlow/Subscribers/DomainEventPublisher.cs
@@ -29,21 +29,25 @@ using System.Threading.Tasks;
 using EventFlow.Aggregates;
 using EventFlow.Core;
 using EventFlow.ReadStores;
+using EventFlow.Sagas;
 
 namespace EventFlow.Subscribers
 {
     public class DomainEventPublisher : IDomainEventPublisher
     {
         private readonly IDispatchToEventSubscribers _dispatchToEventSubscribers;
+        private readonly IDispatchToSagas _dispatchToSagas;
         private readonly IReadOnlyCollection<ISubscribeSynchronousToAll> _subscribeSynchronousToAlls;
         private readonly IReadOnlyCollection<IReadStoreManager> _readStoreManagers;
 
         public DomainEventPublisher(
             IDispatchToEventSubscribers dispatchToEventSubscribers,
+            IDispatchToSagas dispatchToSagas,
             IEnumerable<IReadStoreManager> readStoreManagers,
             IEnumerable<ISubscribeSynchronousToAll> subscribeSynchronousToAlls)
         {
             _dispatchToEventSubscribers = dispatchToEventSubscribers;
+            _dispatchToSagas = dispatchToSagas;
             _subscribeSynchronousToAlls = subscribeSynchronousToAlls.ToList();
             _readStoreManagers = readStoreManagers.ToList();
         }
@@ -67,6 +71,9 @@ namespace EventFlow.Subscribers
 
             // Update subscriptions AFTER read stores have been updated
             await _dispatchToEventSubscribers.DispatchAsync(domainEvents, cancellationToken).ConfigureAwait(false);
+
+            // Update sagas
+            await _dispatchToSagas.ProcessAsync(domainEvents, cancellationToken).ConfigureAwait(false);
         }
     }
 }